requirements.txt
openai-whisper
pandas
pydantic
pydantic
python-dotenv
pytest
groq
